+++ package.json
@@ -23,7 +23,6 @@
     "react-native-icons": "^0.7.1",
     "react-native-picker-select": "^5.2.2",
     "react-native-status-bar-height": "^2.2.0",
-    "react-native-video": "^4.3.1",
     "react-navigation": "^3.0.9",
     "tslib": "1.9.3",
     "tslint": "^5.12.0"
