+++ src/screens/MovieNavigateScreen.tsx
@@ -1,54 +1,85 @@
-import React, { Component } from 'react';
-import { Text, View } from 'react-native';
-import { LinearGradient } from 'expo';
+import * as React from 'react';
+import { 
+  Animated,
+  AppRegistry,
+  Dimensions,
+  FlatList,
+  Image,
+  PanResponder,
+  Platform,
+  ScrollView,
+  StyleSheet,
+  Text,
+  TouchableOpacity,
+  View,
+  WebView,
+} from 'react-native';
 import EStyleSheet from 'react-native-extended-stylesheet';
+import Video from 'react-native-video';
 
-export default class MyComponent extends Component {
+
+// ダミーデータ
+const playlistId = 'PLL4UBL_GFXMlMjpCQdOKEUw_GT7AFB-SX';
+const playlistYoutubeIds = ['uS00Bd9ZVz4', '9iieMHXubJU', '_Hdk2vXESB0', 'X8p0y5KXdIk', 'yYbWPRuxK1U', '33Np_lJseBE', 'z3VW7TNklww', 'tNMatlAOOcs', 'EyT0uC1D1I8', 'bxG4LTsLThE',];
+
+export default class MovieNavigateScreen extends React.Component {
 
   public static navigationOptions = {
     headerStyle: {
       display: 'none',
     },
   };
+  
+  constructor (props) {
+    super(props);
+    this.state = {
+      // cacheはキャッシュしたhtmlを読み込ませないため、playlistIdはyoutubeプレイヤーの初期値を送る
+      youtubeURL: `https://haduki1208-app.firebaseapp.com/youtube.html?cache=${new Date().getTime()}&playlistId=${playlistId}`,
+    };
+  }
+
   public render() {
-    return(
-      <View style={styles.for_section}>
-        <View style={styles.section__header}></View>
-        <View style={styles.section__main_container}>
-          <View style={styles.for_container}>
-            <View style={styles.container__main_content}>
-              <View style={styles.container__movie}></View>
-              <View style={styles.container__option_window}></View>
-              <View style={styles.container__controller}></View>
-            </View>
-            <View style={styles.container__thumbnail}></View>
-          </View>
+    return (
+      <View style={styles.movie_navigate}>
+        <View style={styles.movie_navigate__play}>
+          <Video
+            source={{ uri: 'http://d23dyxeqlo5psv.cloudfront.net/big_buck_bunny.mp4' }}
+            rate={1.0}
+            volume={1.0}
+            isMuted={false}
+            resizeMode="cover"
+            shouldPlay
+            isLooping
+            style={{ width: 300, height: 300 }}
+          />
+        </View>
+        <View style={styles.movie_navigate__controller_wrap}>
+        
+          <TouchableOpacity>
+            <Text style={styles.movie_naviate__controller_play}>再生</Text>
+          </TouchableOpacity>
+          <TouchableOpacity>
+            <Text style={styles.movie_naviate__controller_stop}>停止</Text>
+          </TouchableOpacity>
+          <TouchableOpacity>
+            <Text style={styles.movie_naviate__controller_seekbar}>シークバー</Text>
+          </TouchableOpacity>        
+          <Text style={styles.movie_naviate__controller_seekbar}>再生時間</Text>
         </View>
       </View>
     );
   }
 }
 
-        // <View style={styles.for_container}>
-        // <View style={styles.container__main_content}>
-          // <View style={styles.container__movie}></View>
-          // <View style={styles.container__option_window}></View>
-          // <View style={styles.container__controller}></View>
-        // </View>
-        // <View style={styles.container__thumbnail}></View>
-        // </View>
-
-
-EStyleSheet.build({});
+EStyleSheet.build();
 
 const styles = EStyleSheet.create({
-  for_section: {flex: 1, top: '5%', backgroundColor: 'red',},
-  section__header: {flex: 0.1, backgroundColor: 'blue',},
-  section__main_container: {flex: 0.9, backgroundColor: 'green',},
-  for_container: {flex: 0.8,  width: '100%', justifyContent: 'center', backgroundColor: 'pink',},
-  container__main_content: {position: 'relative', flex: 1, justifyContent: 'center', backgroundColor: 'white',},
-  container__movie: {padding: 10, position: 'absolute', top: 0, zIndex: 10, width: '100%', opacity: 0.2, backgroundColor: 'yellow',},
-  container__option_window: {flex:1, zIndex: 1, width: '100%', backgroundColor: 'brown',},
-  container__controller: {padding: 10, position: 'absolute', bottom: 0, zIndex: 10, width: '100%', opacity: 0.2, backgroundColor: 'gray',},
-  container__thumbnail: {position: 'absolute', bottom: 0, backgroundColor: 'black',},
+  movie_navigate: {
+    marginTop: '20%',
+    width: '100%',
+    height: 'auto',
+    flex: 1,
+    justifyContent: "flex-start",
+  },
 });
+
